Crate alpha_runtime

The Robonomics runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants

A set of constant values used in substrate runtime.

Structs

AdjustmentVariable

AssetDeposit

AssetsStringLimit

AuctionCost

AuctionDuration

BaseFilter

BasicDeposit

BlockHashCount

BlockReward

BondingDuration

BonusReward

DealWithFees

DepositBase

DepositFactor

ExistentialDeposit

FieldDeposit

GenesisConfig

MaxAdditionalFields

MaxLocks

MaxRegistrars

MaxReserves

MaxSignatories

MaxSubAccounts

MaximumMessageSize

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1)

MetadataDepositPerByte

MinimalBid

MinimumMultiplier

MinimumPeriod

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ReferenceCallWeight

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

StakeReward

SubAccountDeposit

TargetBlockFullness

TransactionByteFee

Version

WeightLimit

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

WindowSize

Enums

Call

Event

OriginCaller

Constants

VERSION

Standalone runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Assets

AssetsConfig

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

CumulusXcm

Datalog

DigitalTwin

DmpQueue

Executive

Executive: handles dispatch to the various modules.

Header

Block header type as expected by this runtime.

Identity

Launch

Liability

Lighthouse

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

RWS

SignedExtra

The SignedExtension to the basic transaction logic.

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

XcmpQueue